Remove dialog listener
