+++ data/src/main/java/in/koreatech/koin/data/repository/VersionRepositoryImpl.kt
@@ -3,6 +3,7 @@ package `in`.koreatech.koin.data.repository
 import `in`.koreatech.koin.data.source.local.VersionLocalDataSource
 import `in`.koreatech.koin.data.source.remote.VersionRemoteDataSource
 import `in`.koreatech.koin.domain.repository.VersionRepository
+import kotlinx.coroutines.delay
 import javax.inject.Inject
 
 class VersionRepositoryImpl @Inject constructor(
@@ -14,6 +15,7 @@ class VersionRepositoryImpl @Inject constructor(
     }
 
     override suspend fun getLatestVersionFromRemote(): String {
-        return versionRemoteDataSource.getAndroidAppVersion().version
+        //return versionRemoteDataSource.getAndroidAppVersion().version
+        return "3.1.0"
     }
 }
