+++ koin/src/main/res/layout/version_update_dialog.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
+        android:layout_height="400dp">
 
         <RelativeLayout
             android:id="@+id/version_update_dialog_title_relative_layout"
